@@ -6,7 +6,7 @@ description: Demonstrates how to use the fast_barcode_scanner plugin.
 publish_to: 'none'
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
