@@ -5,15 +5,14 @@ author: Joshua Hoogstraat
 homepage: https://github.com/jhoogstraat/fast_barcode_scanner
 
 environment:
-  sdk: ">=2.8.0 <3.0.0"
-  flutter: ">=1.12.13 < 2.0.0"
+  sdk: '>=2.12.0 <3.0.0'
+  flutter: ">=2.0.0 < 3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
   fast_barcode_scanner_platform_interface:
     path: ../fast_barcode_scanner_platform_interface
-
 
 dev_dependencies:
   flutter_test:
