@@ -7,12 +7,12 @@ author: Joshua Hoogstraat
 version: 0.0.1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
-  flutter: ">=1.10.0 < 2.0.0"
+  sdk: ">=2.12.0 <3.0.0"
+  flutter: ">=2.0.0 < 3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
   fast_barcode_scanner:
     path: ../fast_barcode_scanner
-  soundpool: ^1.1.1
+  soundpool: ^1.1.3
